started to handle the issue where a nodata slice is requested

+++ data_cube_wcs/views.py
@@ -203,9 +203,11 @@ class GetCoverage(View):
         dataset = utils.get_stacked_dataset(dc_parameters, individual_dates, date_ranges)
 
         if 'time' in dataset:
-            dataset = dataset.isel(time=0).astype('float64')
+            dataset = dataset.isel(time=0)
+
         response_mapping = {'GTiff': utils.get_tiff_response, 'netCDF': utils.get_netcdf_response}
         return HttpResponse(
-            response_mapping[coverage_data.cleaned_data['FORMAT']](dataset, coverage_data.cleaned_data['RESPONSE_CRS'],
+            response_mapping[coverage_data.cleaned_data['FORMAT']](dataset.astype('float64'),
+                                                                   coverage_data.cleaned_data['RESPONSE_CRS'],
                                                                    coverage_data.cleaned_data['FORMAT']),
             content_type=forms.AVAILABLE_FORMATS[coverage_data.cleaned_data['FORMAT']])
